+++ controller/Changelog.md
@@ -1,3 +1,9 @@
+# [2023.9.1-VALIDATION] - 2024-03-12
+
+## Changed
+
+- kiosk: Automatically give keyboard focus to active web views
+
 # [2023.9.0] - 2023-09-12
 
 # [2023.9.0-VALIDATION] - 2023-09-11